refactor: split admin cli into typescript modules
