+++ packages/admin/src/blessed.d.ts
@@ -1,0 +1,4 @@
+declare module "blessed" {
+  const blessed: any
+  export = blessed
+}
