+++ packages/admin/src/ui/blessed.ts
@@ -1,0 +1,3 @@
+import * as blessedModule from "blessed"
+
+export const blessed = (blessedModule as any).default ?? blessedModule
